refactor(writer): split writer internals
